refactor monitoring, remove redundant includes
