+++ src/include/CMakeLists.txt
@@ -2,6 +2,7 @@ CMAKE_MINIMUM_REQUIRED(VERSION 3.6)
 
 ADD_SUBDIRECTORY(kvs)
 ADD_SUBDIRECTORY(route)
+ADD_SUBDIRECTORY(monitor)
 ADD_SUBDIRECTORY(lattices)
 ADD_SUBDIRECTORY(utils)
 ADD_SUBDIRECTORY(zmq)
